[ZENDNNL] OMP overhead removal from lowoha operators  (#369)
* [ZENDNNL] OMP api call overhead opt

-ZENAI-3159
-Centralized OMP thread control into a common omp_thread_control.hpp utility
-Optimized OMP API access for matmul, reorder, embedding bag, softmax,
 SDPA, pooling, normalization, and convolution operators
-Removed duplicated thread-limit logic from individual operator headers
-Updated gtests for data type compatibility

+++ zendnnl/src/lowoha_operators/embedding_bag/lowoha_embag_common.hpp
@@ -70,7 +70,7 @@ struct embag_params_t {
   embag_algo_t algo;             // Reduction algorithm (sum/mean/max/none)
   uint64_t num_embeddings;       // Number of rows in the embedding table
   uint64_t embedding_dim;        // Dimension of each embedding vector
-  uint64_t num_threads;          // Number of threads
+  int32_t num_threads;           // Number of threads (0 = auto)
   uint64_t num_indices;          // Total number of indices
   uint64_t num_bags;             // Number of bags (output rows)
   bool is_weights;               // Whether weights are present
